+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Fixed
 
 - Removed explicit line breaks in list item content
+- Removed explicit line breaks in table cells
 
 ## 0.27.0 - 2026-09-16